refactor package structure and use stix naming
